requirements.txt
absl-py==0.11.0
aiohttp==3.7.3
aiohttp-cors==0.7.0
aioredis==1.3.1
apache-beam==2.23.0
apache-flink==1.12.2
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
async-timeout==3.0.1
atari-py==0.2.6
attrs==20.3.0
avro-python3==1.9.1
backcall==0.2.0
bitarray==1.6.1
bleach==3.2.1
blessings==1.7
box2d-py==2.3.8
cached-property==1.5.2
cachetools==4.2.0
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cloudpickle==1.2.2
colorama==0.4.4
colorful==0.5.4
contextvars==2.4
crcmod==1.7
cycler==0.10.0
dataclasses==0.8
decorator @ file:///home/ktietz/src/ci/decorator_1611930055503/work
defusedxml==0.6.0
dgl-cu101==0.6a210203
dill==0.3.1.1
dm-tree==0.1.5
docopt==0.6.2
entrypoints==0.3
et-xmlfile==1.0.1
fastavro==0.23.6
filelock==3.0.12
fire==0.4.0
fsspec==0.8.5
future==0.18.2
gast==0.2.2
google-api-core==1.24.1
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.52.0
gpustat==0.6.0
grpcio==1.34.0
gym==0.18.0
h5py==3.1.0
hdfs==2.6.0
hiredis==1.1.0
httplib2==0.17.4
idna==2.10
idna-ssl==1.1.0
immutables==0.14
importlib-metadata==3.3.0
impyla==0.16.0
iniconfig==1.1.1
ipdb==0.13.7
ipykernel==5.4.2
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.0
jsonpatch==1.32
jsonpickle==1.2
jsonpointer==2.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lightgbm==3.1.1
llvmlite==0.35.0
lz4==3.1.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
mock==2.0.0
msgpack==1.0.2
multidict==5.1.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx @ file:///tmp/build/80754af9/networkx_1598376031484/work
notebook==6.1.6
numba==0.52.0
numpy==1.19.4
nvidia-ml-py3==7.352.0
oauth2client==3.0.0
oauthlib==3.1.0
opencensus==0.7.11
opencensus-context==0.1.2
opencv-python==4.4.0.46
opencv-python-headless==4.3.0.36
openpyxl==3.0.7
opt-einsum==3.3.0
or-gym==0.1.5
ortools==8.1.8487
packaging==20.8
pandas==0.25.3
pandocfilters==1.4.3
parso==0.8.1
pbr==5.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pluggy==0.13.1
ply==3.11
prometheus-client==0.9.0
prompt-toolkit==3.0.9
protobuf==3.14.0
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
py-spy==0.3.3
py4j==0.10.8.1
pyarrow==0.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.2
pyglet==1.5.0
Pygments==2.7.3
pymongo==3.11.3
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.0
python-json-logger==2.0.1
pytorch-lightning==1.1.1
pytz==2020.5
PyYAML==5.3.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
ray==1.1.0
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
rslib==2.2.1
sasl==0.2.1
scikit-learn==0.24.0
scipy==1.5.4
Send2Trash==1.5.0
six @ file:///tmp/build/80754af9/six_1605205335545/work
SQLAlchemy==1.3.22
structlog==20.2.0
tabulate==0.8.7
tensorboard==1.15.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
thrift==0.13.0
thrift-sasl==0.2.1
thriftpy==0.3.9
thriftpy2==0.4.12
toml==0.10.2
torch @ file:///root/dengqilin/reclib/data/torch/torch-1.7.0%2Bcu101-cp36-cp36m-linux_x86_64.whl
torchaudio @ file:///root/dengqilin/reclib/data/torch/torchaudio-0.7.0-cp36-cp36m-linux_x86_64.whl
torchfile==0.1.0
torchnet==0.0.4
torchsummary==1.5.1
torchtext==0.8.0
torchvision @ file:///root/dengqilin/reclib/data/torch/torchvision-0.8.1%2Bcu101-cp36-cp36m-linux_x86_64.whl
tornado==6.1
tqdm==4.54.1
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.2
visdom==0.1.8.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.0.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xgboost==1.3.1
yarl==1.6.3
zipp==3.4.0
